feat: 비즈니스 로직과 DOM event 분리 (#137)
