+++ src/hooks/modal/usePlanModalCategory.ts
@@ -1,4 +1,4 @@
-import { useCallback, useState } from 'react';
+import { useState } from 'react';
 
 import { MAX_CANDIDATE_LENGTH } from '@/constants';
 import { useCategoryQuery } from '@/hooks/query/category';
@@ -56,58 +56,18 @@ const usePlanModalCategory = () => {
     setCategoryInput('');
   };
 
-  const onKeydown = useCallback(
-    (e: React.KeyboardEvent<HTMLInputElement>) => {
-      if (e.key !== 'ArrowDown' && e.key !== 'ArrowUp' && e.key !== 'Enter') {
-        return;
-      }
-
-      if (e.nativeEvent.isComposing) {
-        return;
-      }
-
-      if (filteredCategories.length === 0) {
-        return;
-      }
-
-      e.preventDefault();
-
-      if (e.key === 'Enter') {
-        focusedCategory && setSelectedCategory(focusedCategory.id);
-        return;
-      }
-
-      let index = filteredCategories.findIndex(
-        (category) => category.name === focusedCategory?.name,
-      );
-
-      const lastIndex = filteredCategories.length - 1;
-
-      if (index === -1) {
-        index = index === -1 ? 0 : index;
-      } else if (e.key === 'ArrowDown') {
-        index = index === lastIndex ? 0 : index + 1;
-      } else if (e.key === 'ArrowUp') {
-        index = index === 0 ? lastIndex : index - 1;
-      }
-
-      setFocusedCategory(filteredCategories[index]);
-    },
-    [filteredCategories, focusedCategory],
-  );
-
   return {
     categoryInput,
     filteredType,
     filteredCategories,
     focusedCategory,
     selectedCategory,
+    setFocusedCategory,
     setSelectedCategory,
     setCategoryInput,
     selectCategory,
     filterCategoriesCb,
     clearCategory,
-    onKeydown,
   };
 };
